requirements.txt
asgiref==3.8.1
Django==4.2.20
django-ckeditor==6.7.2
django-js-asset==2.2.0
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
